requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
python-multipart>=0.0.9
typing-extensions>=4.9.0
python-dotenv>=1.0.1
httpx>=0.26.0
pymongo
pymupdf
sentence-transformers
langchain>=0.1.0
google-generativeai
pinecone
